+++ 0x1A-hash_tables/6-hash_table_delete.c
@@ -2,7 +2,7 @@
 
 void hash_table_delete(hash_table_t *ht)
 {
-    int i = 0;
+    u_long i = 0;
     hash_node_t *tmp = NULL;
     
     if(ht)
